+++ src/main/java/lotto/domain/model/generator/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model.generator;
 
 import java.util.List;
