+++ src/main/java/lotto/domain/model/lotto/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model.lotto;
 
 import lotto.constants.LottoConfig;
