+++ src/main/java/lotto/domain/model/lotto/LottoWinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model.lotto;
 
 public class LottoWinningNumbers {
