+++ src/main/java/lotto/domain/model/result/WinningRankCounts.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model.result;
 
 import java.util.Arrays;
 import java.util.EnumMap;
