+++ src/main/java/lotto/view/InputProcessor.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.List;
