+++ src/test/java/lotto/PaidAmountTest.java
@@ -1,11 +1,9 @@
 package lotto;
 
+import lotto.domain.model.lotto.PaidAmount;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
