+++ src/test/java/lotto/WinningRankCountsTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.model.result.WinningRank;
+import lotto.domain.model.result.WinningRankCounts;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
