+++ src/test/java/lotto/WinningRankTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
+import lotto.domain.model.result.WinningRank;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
